tests/wpt/tests/credential-management/fedcm-multi-idp/get-before-onload-and-during-dom-content-loaded.https.html

<!DOCTYPE html>
<title>Federated Credential Management API multi IDP get before onload and during DOMContentLoaded test.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>

<script type="module">
import {set_fedcm_cookie,
        set_alt_fedcm_cookie,
        request_options_with_mediation_required,
        alt_request_options_with_mediation_required,
        fedcm_select_account_promise} from '../support/fedcm-helper.sub.js';

let cookies_promise = Promise.all([set_fedcm_cookie(), set_alt_fedcm_cookie()]);

promise_test(async t => {
  let has_dom_content_loaded = false;
  let rejection;
  const dom_content_loaded = new Promise(resolve => {
      document.addEventListener('DOMContentLoaded', async () => {
          const second_cred = navigator.credentials.get(alt_request_options_with_mediation_required());
          rejection = promise_rejects_dom(t, 'NetworkError', second_cred);
          has_dom_content_loaded = true;
          resolve();
      });
  });
  assert_false(has_dom_content_loaded);
  const first_cred = navigator.credentials.get(request_options_with_mediation_required());
  await Promise.all([cookies_promise, dom_content_loaded]);
  assert_true(has_dom_content_loaded);

  await fedcm_select_account_promise(t, 0);
  const first = await first_cred;
  assert_equals(first.token, "token");
  return rejection;
}, "A `get` call before onload and a `get` call during DOMContentLoaded event should combine despite being called from different tasks.");

</script>
